+++ cli.py
@@ -1,23 +1,47 @@
 import argparse
-from src.backends import poetrade, poeofficial
+from typing import List, Dict, Any, Set
+
+from src.config.user_config import UserConfig
+from src.core.backends import poetrade, poeofficial
 from src.pathfinder import PathFinder
 from src.commons import league_names
-from src.assets import ItemList
+from src.trading import ItemList
 
 default_backend = poeofficial
 item_list = ItemList.load_from_file()
 
 
 def log_conversions(conversions, league, currency, limit):
-    for c in conversions[currency][:limit]:
+
+    unique_conversions = get_independent_conversions(conversions[currency], limit)
+
+    for c in unique_conversions[:limit]:
         log_conversion(c, league)
 
 
+def get_independent_conversions(conversions: List[Dict[str, Any]], limit: int) -> List[Dict]:
+    seen_traders: Set[str] = set()
+    unique_conversions = []
+
+    for conversion in conversions:
+        trader_names = [t["contact_ign"] for t in conversion["transactions"]]
+        has_seen_trader = any([True for x in trader_names if x in seen_traders])
+        if has_seen_trader:
+            continue
+
+        for t in trader_names:
+            seen_traders.add(t)
+
+        unique_conversions.append(conversion)
+
+    return unique_conversions
+
+
 def log_conversion(c, league):
-    print("{} {} -> {} {}: {} {}".format(c["starting"], c["from"], c["ending"],
-                                         c["to"], c["winnings"], c["to"]))
+    print("\t{} {} -> {} {}: {} {}".format(c["starting"], c["from"], c["ending"],
+                                           c["to"], c["winnings"], c["to"]))
     for t in c["transactions"]:
-        print("\t@{} Hi, I'd like to buy your {} {} for {} {} in {}. ({}x)".format(
+        print("\t\t@{} Hi, I'd like to buy your {} {} for {} {} in {}. ({}x)".format(
             t["contact_ign"],
             t["received"],
             t["to"],
@@ -41,15 +65,15 @@ parser.add_argument(
 parser.add_argument(
     "--currency",
     default="all",
-    #choices=cli_default_items,
+    # choices=cli_default_items,
     type=str,
     help="Full name of currency to flip, ie. 'Cartographer's Chisel, or 'Chaos Orb'. Defaults to all currencies.",
 )
 parser.add_argument(
     "--limit",
-    default=3,
+    default=5,
     type=int,
-    help="Limit the number of displayed conversions. Defaults to 3.",
+    help="Limit the number of displayed conversions. Defaults to 5.",
 )
 parser.add_argument(
     "--fullbulk",
@@ -79,14 +103,18 @@ config = {"fullbulk": fullbulk}
 chosen_currencies = item_list.get_item_list_for_backend(backend, config)
 
 # Load excluded trader list
-with open("excluded_traders.txt", "r", encoding='utf-8') as f:
-    excluded_traders = [x.strip() for x in f.readlines()]
+with open("config/excluded_traders.txt", "r", encoding='utf-8') as f:
+    excluded_traders = [x.strip().lower() for x in f.readlines()]
 
-p = PathFinder(league, chosen_currencies, backend, excluded_traders,
+# Load user config
+user_config = UserConfig.from_file()
+
+p = PathFinder(league, chosen_currencies, backend, user_config, excluded_traders,
                use_filter)
-p.run(3, True)
+p.run(3)
 
 try:
+    print("\n")
     if currency == "all":
         for c in p.graph.keys():
             log_conversions(p.results, league, c, limit)